+++ Cineaste/Extension/UIViewController+ShareMovie.swift
@@ -9,7 +9,42 @@
 import UIKit
 
 extension UIViewController {
-    func share(movie: Movie) {
+    /// This presents an `UIActivityViewController` to share a movie.
+    /// - Parameters:
+    ///   - movie: The movie which should be shared.
+    ///   - barButtonItem: The barButtonItem where the `UIActivityViewController`
+    ///   should be presented on (for iPads).
+    func share(movie: Movie, onBarButtonItem barButtonItem: UIBarButtonItem) {
+        let activityController = getConfiguredActivityVC(for: movie)
+        activityController.popoverPresentationController?.barButtonItem = barButtonItem
+
+        DispatchQueue.main.async {
+            self.present(activityController, animated: true)
+        }
+    }
+
+    /// This presents an `UIActivityViewController` to share a movie. 
+    /// - Parameters:
+    ///   - movie: The movie which should be shared.
+    ///   - sourceView: The source view where the `UIActivityViewController`
+    ///   should be presented on (for iPads).
+    func share(movie: Movie, onSourceView sourceView: UIView) {
+        let activityController = getConfiguredActivityVC(for: movie)
+
+        activityController.popoverPresentationController?.sourceView = sourceView
+        activityController.popoverPresentationController?.sourceRect = CGRect(
+            x: sourceView.bounds.midX,
+            y: sourceView.bounds.midY,
+            width: 0,
+            height: 0
+        )
+
+        DispatchQueue.main.async {
+            self.present(activityController, animated: true)
+        }
+    }
+
+    private func getConfiguredActivityVC(for movie: Movie) -> UIActivityViewController {
         var items = [Any]()
 
         items.append(movie.title)
@@ -18,13 +53,9 @@ extension UIViewController {
             items.append(url)
         }
 
-        let activityController = UIActivityViewController(
+        return UIActivityViewController(
             activityItems: items,
             applicationActivities: nil
         )
-
-        DispatchQueue.main.async {
-            self.present(activityController, animated: true)
-        }
     }
 }